ui tweaks -- in-game ui stuff: slightly better handling of meme levels: - "84-0" literal replaced with relevant define - the check is now shared between patches
i thought that all the question marks i added last time is enough to handle these and tried to remove the uid checks. turns out, (at least for museyao), the bars are present, but something else inside them is not; i should really look into the meme ui, but for now we'll keep the uid checking -- it's not like it's april fools every month, is it? (^_−)