ui tweaks: fixed the title for the longest album being sligthly cut off on its ends by gradient/trasnparency masks
this is committed as is as soon i got it working, now to clean up the "prototype" code somewhat

basically, this shrinks the mask images in half so they no longer overlap with the "MD-level... ...Blu-ray" text. Technically, we can just not render them, but there are also some solid violet coloring on the edges of the entire area, into which the masks merge, and it looks scuffed without them, so they are merely shrunk instead. The right one is also moved to keep its position, because resizing keeps orininal top and left coordinates